[phi] Refactor and make phi directories auto create on run
