+++ common/log.go
@@ -1,0 +1,7 @@
+package common
+
+import (
+	"log"
+)
+
+func DefaultLogSettings() { log.SetFlags(log.Ldate | log.Ltime | log.Lshortfile) }
